+++ packages/react-relay/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.1.12...@wundergraph/react-relay@0.1.13) (2023-05-03)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
+## [0.1.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.1.11...@wundergraph/react-relay@0.1.12) (2023-05-03)
+
+**Note:** Version bump only for package @wundergraph/react-relay
+
 ## [0.1.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/react-relay@0.1.10...@wundergraph/react-relay@0.1.11) (2023-04-27)
 
 **Note:** Version bump only for package @wundergraph/react-relay